+++ build.gradle
@@ -1,7 +1,7 @@
 plugins {
     id 'java'
     id 'application'
-//    id 'checkstyle'
+    id 'checkstyle'
     id 'com.github.johnrengelman.shadow' version '7.1.2'
 }
 
@@ -52,12 +52,12 @@ shadowJar {
     archiveClassifier = null
 }
 
-run{
+run {
     standardInput = System.in
 }
 
-//checkstyle {
-//    toolVersion = '11.0.0'
-//}
+checkstyle {
+    toolVersion = '11.0.0'
+}
