+++ src/main/java/Listgp/CLI.java
@@ -1,4 +1,4 @@
-package Listgpt;
+package Listgp;
 
 import java.util.Scanner;
