+++ src/main/java/Listgp/Deadline.java
@@ -1,4 +1,4 @@
-package Listgpt;
+package Listgp;
 
 import java.time.DateTimeException;
 import java.time.LocalDate;
