+++ src/main/java/Listgp/DialogBox.java
@@ -1,4 +1,4 @@
-package Listgpt;
+package Listgp;
 
 import javafx.geometry.Pos;
 import javafx.geometry.Insets;
