+++ src/main/java/Listgp/Launcher.java
@@ -1,4 +1,4 @@
-package Listgpt;
+package Listgp;
 
 import javafx.application.Application;
