+++ src/main/java/Listgp/Parser.java
@@ -1,4 +1,4 @@
-package Listgpt;
+package Listgp;
 
 import java.time.DateTimeException;
 import java.util.regex.Matcher;
