+++ src/main/java/Listgp/Task.java
@@ -1,6 +1,4 @@
-package Listgpt;
-
-import java.util.Optional;
+package Listgp;
 
 /**
  * Represents a task with a name and a completion state.
