+++ src/test/java/Listgp/DeadlineTest.java
@@ -1,4 +1,4 @@
-package Listgpt;
+package Listgp;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
