+++ src/test/java/Listgp/ParserTest.java
@@ -1,4 +1,4 @@
-package Listgpt;
+package Listgp;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import org.junit.jupiter.api.Test;
