+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
    __________________________________
-   Hello! I'm Listgpt!
+   Hello! I'm Listgp!
    What can I do for you?
    __________________________________
